Moving sbt model classes to their own package.
